+++ examples/redis/processor.php
@@ -9,7 +9,7 @@ require_once __DIR__.'/../../vendor/autoload.php';
 
 $redis = new Redis();
 $redis->connect('127.0.0.1', 6379);
-$driver = new RedisSetDriver($redis);
+$driver = new RedisSetDriver($redis, 'hermes');
 
 $dispatcher = new Dispatcher($driver);