+++ tests/Driver/CustomSerializerTest.php
@@ -3,13 +3,13 @@ declare(strict_types=1);
 
 namespace Tomaj\Hermes\Test;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Tomaj\Hermes\Test\Driver\DummyDriver;
 use Tomaj\Hermes\Test\Driver\DummySerializer;
 use Tomaj\Hermes\MessageInterface;
 use Tomaj\Hermes\Message;
 
-class CustomSerializerTest extends PHPUnit_Framework_TestCase
+class CustomSerializerTest extends TestCase
 {
     public function testCustomSerializerTest()
     {