+++ tests/Restart/DispatcherRestartTest.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Tomaj\Hermes\Test;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Tomaj\Hermes\Test\Driver\DummyDriver;
 use Tomaj\Hermes\Test\Handler\TestHandler;
 use Tomaj\Hermes\Message;
 use Tomaj\Hermes\Dispatcher;
 use Tomaj\Hermes\Test\Restart\StopRestart;
 
-class HandleRestartTest extends PHPUnit_Framework_TestCase
+class HandleRestartTest extends TestCase
 {
     public function testEmitWithDummyDriverNoRestart()
     {